xrpl-cosmos/requirements.txt
requests==2.32.5
bech32==1.2.0
eth-account==0.13.7
python-dotenv==1.2.1
